# services/order/openapi.yaml
openapi: 3.0.1
info:
  title: Order Service API
  version: 1.0.0
paths:
  /orders:
    get:
      summary: Get all orders
      responses:
        '200':
          description: List of orders
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                status:
                  type: string
              required:
                - userId
                - items
      responses:
        '201':
          description: Order created
        '400':
          description: Invalid data
  /orders/{id}:
    get:
      summary: Get an order by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Order found
        '404':
          description: Order not found
    put:
      summary: Update an order by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
      responses:
        '200':
          description: Order updated
        '404':
          description: Order not found
    delete:
      summary: Delete an order by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Order deleted
        '404':
          description: Order not found
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy